videopipe-viz/requirements.txt
moviepy==1.0.3
numpy==1.23.3
pandas==1.5.1
Pillow==9.3.0
Pillow==9.5.0
